// src/app/admin-view/models/localbody.ts
export class LocalBody {
    AssemblyId: string;
    AssemblyName: string;
    DistId: number;
    DistName: string;
    DivisionId: string;
    Id: string;
    LocalBodyId: string;
    LocalBodyName: string;
    LocalBodyType: string;
    LocalBodyTypeId: number;
    // localBodyTypeText: string;
    MobileNo: string;
    SECElectionId: string;
    StateId: number;
    StateName: string;
    ReportId: string;
    TalukaId: number;
    TalukaName: string;
    ACNo: string;
    division: string;
    TokenId: string;
    Status: string;
    checkedStatus: boolean;
    lbStatus: boolean ;
}